refactor agents and generation module
