+++ server/src/modules/agents/build.agent.ts
@@ -12,7 +12,7 @@ import {
     isFrontendWorkspacePath,
 } from '../generation/generation.utils'
 import { readChatCompletionText, retryAsync } from './agents.utils'
-import { BUILD_AGENT_PROMPT } from './build.prompt'
+import { BUILD_AGENT_PROMPT, BUILD_PATCH_AGENT_PROMPT } from './build.prompt'
 
 type ProjectIntent = z.infer<typeof projectIntentSchema>
 type ProjectPlan = z.infer<typeof projectPlanSchema>
@@ -204,34 +204,6 @@ const selectPatchContext = (targetPath: string, currentFiles: Record<string, str
             content: currentFiles[path],
         }))
 }
-
-const BUILD_PATCH_AGENT_PROMPT = `You are the Build Agent for a December follow-up patch.
-
-Generate EXACTLY ONE frontend file for one patch operation.
-Return ONLY raw file content.
-No markdown.
-No code fences.
-No JSON.
-No filename.
-No explanation.
-
-You are given:
-- the user's edit or runtime-fix request
-- one planned patch operation
-- the current target file content when it exists
-- a small set of related project files
-
-Rules:
-- Produce the complete final content for the requested path only.
-- Keep the change targeted to the operation instructions.
-- Do not redesign unrelated UI or rewrite the whole project style.
-- Do not add imports for files that are not present or clearly planned by the operation.
-- Stay browser-only Bun React TypeScript.
-- Do not write backend, API, database, env, Docker, or server code.
-- Never output TODO, FIXME, placeholders, pseudocode, or markdown fences.
-- package.json and tsconfig.json must remain valid JSON.
-- React files must have valid imports and exports.
-- If fixing an error, preserve working behavior and fix the smallest likely cause.`
 
 export const generateProjectPatchFile = async (
     data: GenerateProjectPatchInput & { model?: string }
